# gsmapplications-frontend/swagger/gsm-application.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "GSMApplication API",
    "description": "Microservicio de aplicación con soporte tenant database-per-tenant.",
    "version": "v1"
  },
  "paths": {
    "/api/v1/Application/menu": {
      "get": {
        "tags": [
          "Application"
        ],
        "operationId": "GetMenu",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/GetMenuDTOApiResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetMenuDTOApiResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetMenuDTOApiResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/GetMenuDTOApiResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetMenuDTOApiResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetMenuDTOApiResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/Application/mediaResources": {
      "get": {
        "tags": [
          "Application"
        ],
        "operationId": "GetMediaResources",
        "parameters": [
          {
            "name": "categories",
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/MultimediaResourceDtoListApiResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MultimediaResourceDtoListApiResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/MultimediaResourceDtoListApiResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/MultimediaResourceDtoApiResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MultimediaResourceDtoApiResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/MultimediaResourceDtoApiResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/MultimediaResourceDtoApiResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MultimediaResourceDtoApiResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/MultimediaResourceDtoApiResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/Application/filtered-locations": {
      "get": {
        "tags": [
          "Application"
        ],
        "operationId": "GetFilteredLocations",
        "parameters": [
          {
            "name": "IdLocation",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "CodeLocation",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/LocationDTOListApiResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LocationDTOListApiResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/LocationDTOListApiResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorType": {
        "enum": [
          "Validation",
          "Unauthorized",
          "NotFound",
          "BadRequest",
          "Conflict",
          "Internal",
          "Forbidden"
        ],
        "type": "string"
      },
      "GetMenuDTO": {
        "type": "object",
        "properties": {
          "idProfile": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "menu": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "GetMenuDTOApiResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string",
            "nullable": true
          },
          "data": {
            "$ref": "#/components/schemas/GetMenuDTO"
          },
          "errorType": {
            "$ref": "#/components/schemas/ErrorType"
          },
          "traceId": {
            "type": "string",
            "nullable": true
          },
          "details": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "LocationDTO": {
        "required": [
          "idLocation"
        ],
        "type": "object",
        "properties": {
          "idLocation": {
            "type": "integer",
            "format": "int32"
          },
          "idThirdLocation": {
            "type": "string",
            "nullable": true
          },
          "codeLocation": {
            "type": "string",
            "nullable": true
          },
          "descr": {
            "type": "string",
            "nullable": true
          },
          "email": {
            "type": "string",
            "nullable": true
          },
          "phoneNumber": {
            "type": "string",
            "nullable": true
          },
          "addressLocation": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "LocationDTOListApiResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string",
            "nullable": true
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LocationDTO"
            },
            "nullable": true
          },
          "errorType": {
            "$ref": "#/components/schemas/ErrorType"
          },
          "traceId": {
            "type": "string",
            "nullable": true
          },
          "details": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "MultimediaResourceDto": {
        "type": "object",
        "properties": {
          "resourceCategory": {
            "type": "string",
            "nullable": true
          },
          "resourceOrder": {
            "type": "integer",
            "format": "int32"
          },
          "config": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "MultimediaResourceDtoApiResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string",
            "nullable": true
          },
          "data": {
            "$ref": "#/components/schemas/MultimediaResourceDto"
          },
          "errorType": {
            "$ref": "#/components/schemas/ErrorType"
          },
          "traceId": {
            "type": "string",
            "nullable": true
          },
          "details": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "MultimediaResourceDtoListApiResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string",
            "nullable": true
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MultimediaResourceDto"
            },
            "nullable": true
          },
          "errorType": {
            "$ref": "#/components/schemas/ErrorType"
          },
          "traceId": {
            "type": "string",
            "nullable": true
          },
          "details": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "JWT Authorization header usando el esquema Bearer. Ejemplo: \"Bearer {token}\"",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}